requirements.txt
pandas==0.20.1
numpy==1.13.3
Pillow==5.0.0
kafka-python==1.3.3
pytest==3.3.2
pydicom